Implement server group/cluster providers

+++ clouddriver-azure/src/main/groovy/com/netflix/spinnaker/clouddriver/azure/resources/common/AzureResourceOpsDescription.groovy
@@ -28,7 +28,7 @@ class AzureResourceOpsDescription {
   AzureCredentials credentials
   String region
   String user
-  Integer createdTime
+  Long createdTime
   long lastReadTime
   Map<String,String> tags = [:]
 }
